Clean up the code as per review comments mentioned in PR
