+++ src/main/java/com/ui/LoginReportPortalPage.java
@@ -6,4 +6,6 @@ public class LoginReportPortalPage extends BasePage {
     public LoginReportPortalPage(WebDriver driver) {
         super(driver);
     }
+
+    ///Report portal Login Page Locators and respective page methods will be coded here///
 }
