+++ src/test/java/components/ui/BaseTest.java
@@ -9,6 +9,9 @@ import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.edge.EdgeOptions;
 import org.openqa.selenium.remote.CapabilityType;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testng.annotations.BeforeTest;
 import utilities.ReadPropertyFile;
 
 import java.io.File;
@@ -16,39 +19,21 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import static org.openqa.selenium.remote.CapabilityType.PROXY;
-
+///This class provide the initialization of base set up methods like Weddriver setup before the test triggers///
 public class BaseTest  {
     ReadPropertyFile fr = new ReadPropertyFile();
-    private WebDriver driver;
+    protected WebDriver driver;
+    private final static Logger logger = LoggerFactory.getLogger(BasePage.class);
 
-
-    public WebDriver setUp() throws IOException {
+///This method will set up the driver and return it///
+    public WebDriver getDriver() throws IOException {
 
         String browser =  fr.getPropertyValue("browser");
 
         Path resourceDirectory = Paths.get("src","test","resources");
         String absolutePath = resourceDirectory.toFile().getAbsolutePath();
 
-        if (browser.equalsIgnoreCase("edge"))
-        {
-            try {
-                String[] pathNames = {absolutePath, "\\drivers\\msedgedriver.exe" };
-                String path = String.join(File.pathSeparator, pathNames);
-                System.setProperty("webdriver.edge.driver",path.replaceAll(";",""));
-                EdgeOptions options = new EdgeOptions();
-                options.setCapability(CapabilityType.ACCEPT_INSECURE_CERTS, true);
-                options.addArguments("--remote-allow-origins=*");
-                options.setCapability("ignore-certificate-errors", true);
-
-
-                driver = new EdgeDriver(options);
-            }
-            catch (Exception exp){
-                exp.printStackTrace();
-            }
-        }
-        else if(browser.equalsIgnoreCase("chrome")) {
+        if(browser.equalsIgnoreCase("chrome")) {
             String[] pathNames = {absolutePath, "\\drivers\\chromedriver.exe" };
             String path = String.join(File.pathSeparator, pathNames);
             System.out.println(path);
@@ -63,5 +48,16 @@ public class BaseTest  {
         }
         return driver;
     }
+    /// Close the Browser///
+    public void driverClose()
+    {
+        driver.close();
+        logger.info("Webdriver Closed successfully");
 
+    }
+///Initialize the Webdriver///
+    @BeforeTest
+    public void driverSetup() throws IOException {
+        this.driver = getDriver();
+    }
 }
