+++ src/test/java/components/ui/LoginReportPortalTest.java
@@ -1,23 +1,25 @@
 package components.ui;
 
 import com.ui.LoginReportPortalPage;
-import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
-import utilities.ReadPropertyFile;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.Test;
+
 
 import java.io.IOException;
-
+///This class provide the different Report Portal service test using TestNG annotations
 public class LoginReportPortalTest  extends BaseTest {
+    LoginReportPortalPage loginReportPortalPage;
 
-    ReadPropertyFile fr = new ReadPropertyFile();
-
-
-
-
-    public void openTheReportUrl(WebDriver driver) throws IOException {
-        LoginReportPortalPage loginReportPortalPage = new LoginReportPortalPage(driver);
+    @Test
+    public void testOpenTheReportUrl() throws IOException {
+        loginReportPortalPage = new LoginReportPortalPage(this.driver);
         loginReportPortalPage.navigateToUrl(fr.getPropertyValue("reportPortal"));
         Assert.assertEquals(loginReportPortalPage.getPageTitle(), "Report Portal");
 
+    }
+    @AfterTest
+    public void cleanUp(){
+       driverClose();
     }
 }
