+++ src/test/java/runners/TestNGRunnerTest.java
@@ -12,25 +12,5 @@ import java.io.IOException;
 
 public class TestNGRunnerTest extends LoginReportPortalTest {
 
-    WebDriver driver;
-    LoginReportPortalPage loginReportPortalPage;
-
-    @BeforeTest
-    public void driverSetup() throws IOException {
-        this.driver = setUp();
-    }
-
-    @Test
-    public void testOpenTheReportUrl() throws IOException {
-
-        openTheReportUrl(driver);
-    }
-
-    @AfterTest
-    public void cleanUp(){
-        loginReportPortalPage = new LoginReportPortalPage(driver);
-        loginReportPortalPage.driverClose();
-    }
-
 
 }
